+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### [1.0.0] - UNRELEASED
+
+#### Added
+
+- [📌 #1](https://github.com/CardanoSolutions/kupo/issues/1) - New API endpoint to get application's health, 
+  - `GET v1/health` → [🕮  API Reference](https://cardanosolutions.github.io/kupo/#operation/getHealth)
+
+- [📌 #1](https://github.com/CardanoSolutions/kupo/issues/1) - New command `healthcheck` to perform a health check against a running server. Handy when combined with Docker's HEALTHCHECK feature.
+
+#### Changed
+
+- [📌 #2](https://github.com/CardanoSolutions/kupo/issues/2) - The command-line is now more idempotent, restarting the server with the same options will no longer fail and simply resume syncing.
+
 ### [1.0.0-beta] - 2022-02-27
 
 #### Added
@@ -10,10 +23,10 @@
   - Multiple patterns via the command-line (`--match`)
 
   - Basic API access:
-    - `GET v1/matches`
-    - `GET v1/matches/{pattern-fragment}`
-    - `GET v1/matches/{pattern-fragment}/{pattern-fragment}`
-    - `GET v1/checkoints`
+    - `GET v1/matches` → [🕮  API Reference](https://cardanosolutions.github.io/kupo/#operation/getAllMatches)
+    - `GET v1/matches/{pattern-fragment}` → [🕮  API Reference](https://cardanosolutions.github.io/kupo/#operation/getMatchesUnary)
+    - `GET v1/matches/{pattern-fragment}/{pattern-fragment}` → [🕮  API Reference](https://cardanosolutions.github.io/kupo/#operation/getMatches2Ary)
+    - `GET v1/checkoints` → [🕮  API Reference](https://cardanosolutions.github.io/kupo/#operation/getCheckpoints)
 
   - Multi-level (basic) structured component logging (`--log-level`, `log-level-{component}`)